New CVS commit notification for tcptrace.
If you edit the tcptrace CVS repository, when you check in your changes your
log message will be mailed to the tcptrace-maintainers mailing list.  This
should help keep everyone synced on what is going on; eventually there may be
a more fine-grained method of syncing (i.e. what if we don't want
notifications during periods of heavy modification...), but for now EVERY
commit is going out.

A side effect of this is that you need to make sure your log messages are
clear, detailed, and something you wouldn't mind your mother reading.  ;-)

This is an example of such a log message.
Ethan