Added a small patch to change the output of the "min win adv" field of 'tcptrace -l' output.
Till now, on a window scaled connection, the "min win adv"
field would always be 65535 or whatever window size
was advertised in SYN packet, as SYN packets should not
be scaled according to RFC 1323.

This fix, lets tcptrace ignore the window size advertised
in the SYN packet as being counted as "min win adv" in case
of a window scaled connection.

Added a flag "window_stats_updated_for_scaling"
in the tcb structure in tcptrace.h, to notice the first
window-scaled packet in either direction.
Also modified trace.c to do the required
work of resetting win_min upon seeing such a packet.