Fixed the macros looking for ECN flags : CWR and ECE in the TCP header.
The current macros were looking for those flags in the 6th and 7th
bits of the 32 bit word in TCP header containing the TCP header length
(with the most significant bit counted as the 0th bit).

According to RFC 2481 and draft-ietf-tsvwg-ecn-04.txt,
these bits are actually in the 8th and 9th bits of this 32bit
word.