Updating the following "long" files in the OUTPUTbench directory to reflect the correct values in the "min win adv" field.
These files are interesting in the sense that their minimum
advertised window is in their SYN packets as their window
size keeps increasing during the life time of the connection.
They were advertising a WINDOW SCALE of 0 in their SYN packets
and tcptrace was displaying wrong "min win adv" stats in these
circumstances.