Fixing the bug found when tcptrace was asked to process packets directly from stdin, as in :
tcpdump -n -w - | tcptrace -n stdin

tcptrace exits prematurely with "PCAP error - truncated file" or a similar
error.

This was found to be due to FILE synchronization issues exhibited when
parent and child processes have a FILE pointer open for stdin, which was
pointing to the temporary file in which the first 8k (COMP_HDR_SIZE) chunk
of data is stored to determine the dumpfile type.

The fix was trivial. In the child process (PipeFitting() function in
compress.c), we close stdin, and reopen a FILE pointer to the temporary file.
In the parent process, we purge off the contents of stdin, and point the
read end of the pipe to file descriptor 0. This kind of resets things and
clears up the buffers, and makes the world a better place.