Fixed the bug in the calculation of the "avg win adv" field in long output. Now, we have the nice property that avg. falls in between min. and max. :-)
The win_tot field in the tcb structure that sums up all the window
advertisements seen to calculate the average in the end, was a u_long which
had to be changed to u_llong.

Further, the semantics of what tcptrace reports as window stats (min,max,avg)
when both end-points agree on using window scaling has been modified.
When we know that both sides are infact doing window scaling (we saw both the
SYNs carrying window scale option), we flush out the win_min, win_max,
win_tot stats gathered from the SYN packets so that the min, max, and
avg values reported are only for the scaled window advertisements seen in
the connection.