Fixed the #ifndef __WIN32 to #ifdef __WIN32 for the in6_addr structure definition in ipv6.h
Applied changes to the HTTP module (mod_http.c) based on Yufei Wang's
patch. Now we are more robust to print information on HTTP connections
that get closed with RST segments. We also print "Server Rst Time"/
"Server Last Time" to print the times
of last segments seen in the respective direction when connections
were closed with RST / if we didn't see the connection close (We used to
print only "Server Fin Time" before). Similar information is also printed
for the Client too.